@@ -20,7 +20,7 @@
     "@uniswap/sdk-core": "^4.0.1",
     "@uniswap/v2-sdk": "^3.0.1",
     "daisyui": "^3.5.1",
-    "next": "^13.1.6",
+    "next": "13.4",
     "nextjs-progressbar": "^0.0.16",
     "qrcode.react": "^3.1.0",
     "react": "^18.2.0",
