@@ -1,17 +1,20 @@
+'use client'
 import { useEffect, useState } from "react";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import type { NextPage } from "next";
 import { Transaction, TransactionReceipt, formatEther, formatUnits } from "viem";
 import { usePublicClient } from "wagmi";
 import { hardhat } from "wagmi/chains";
 import { Address } from "~~/components/scaffold-eth";
 import { decodeTransactionData, getFunctionDetails, getTargetNetwork } from "~~/utils/scaffold-eth";
-
-const TransactionPage: NextPage = () => {
+type PageProps = {
+  params: {txHash?: `0x${string}`};
+ 
+};
+const TransactionPage:NextPage<PageProps> = ({params}:PageProps) => {
   const client = usePublicClient({ chainId: hardhat.id });
-
+  const txHash=params?.txHash as `0x${string}`;
   const router = useRouter();
-  const { txHash } = router.query as { txHash?: `0x${string}` };
   const [transaction, setTransaction] = useState<Transaction>();
   const [receipt, setReceipt] = useState<TransactionReceipt>();
   const [functionCalled, setFunctionCalled] = useState<string>();
