@@ -6,7 +6,7 @@ import { MetaHeader } from "~~/components/MetaHeader";
 const Home: NextPage = () => {
   return (
     <>
-      <MetaHeader />
+      {/* <MetaHeader /> */}
       <div className="flex items-center flex-col flex-grow pt-10">
         <div className="px-5">
           <h1 className="text-center mb-8">
